refactor(core): remove deck.gl dependencies
