+++ showcase/index.js
@@ -20,7 +20,7 @@
 
 import ComplexChart from './plot/complex-chart';
 import LineChart from './plot/line-chart';
-import LineChartGL from './plot/line-chart-gl';
+import LineChartCanvas from './plot/line-chart-canvas';
 import LineChartWithStyle from './plot/line-chart-with-style';
 import LineMarkChart from './plot/linemark-chart';
 import BarChart from './plot/bar-chart';
@@ -32,7 +32,7 @@ import AreaChart from './plot/area-chart';
 import AreaChartElevated from './plot/area-chart-elevated';
 import ScatterplotChart from './plot/scatterplot';
 import FauxScatterplotChart from './plot/faux-radial-scatterplot';
-import ScatterplotGLChart from './plot/scatterplot-gl';
+import ScatterplotCanvas from './plot/scatterplot-canvas';
 import HeatmapChart from './plot/heatmap-chart';
 import WidthHeightMarginChart from './plot/width-height-margin';
 import CustomScales from './plot/custom-scales';
@@ -83,7 +83,7 @@ import TreemapExample from './treemap/dynamic-treemap';
 export const showCase = {
   ComplexChart,
   LineChart,
-  LineChartGL,
+  LineChartCanvas,
   LineChartWithStyle,
   LineMarkChart,
   BarChart,
@@ -96,7 +96,7 @@ export const showCase = {
   AreaChartElevated,
   FauxScatterplotChart,
   ScatterplotChart,
-  ScatterplotGLChart,
+  ScatterplotCanvas,
   HeatmapChart,
   WidthHeightMarginChart,
   CustomScales,
