+++ showcase/plot/line-chart-canvas.js
@@ -19,83 +19,94 @@
 // THE SOFTWARE.
 
 import React from 'react';
-import ShowcaseButton from '../showcase-components/showcase-button';
 
+import ShowcaseButton from '../showcase-components/showcase-button';
 import {
   XYPlot,
   XAxis,
   YAxis,
   VerticalGridLines,
   HorizontalGridLines,
-  MarkSeries,
-  MarkSeriesGL,
-  MarkSeriesCanvas,
-  Hint
+  LineMarkSeriesCanvas,
+  LineMarkSeries,
+  LineSeriesCanvas,
+  LineSeries,
+  Crosshair
 } from 'index';
 
 function getRandomData() {
-  return (new Array(100)).fill(0).map(row => ({
-    x: Math.random() * 10,
+  return (new Array(1000)).fill(0).map((row, i) => ({
+    x: i,
     y: Math.random() * 20,
-    size: Math.random() * 10,
-    color: Math.random() * 10,
-    opacity: Math.random() * 0.5 + 0.5
+    color: Math.random() * 10
   }));
 }
+
+const randomData = getRandomData();
+
 const colorRanges = {
   typeA: ['#59E4EC', '#0D676C'],
   typeB: ['#EFC1E3', '#B52F93']
 };
 
-const randomData = getRandomData();
 const nextType = {
   typeA: 'typeB',
   typeB: 'typeA'
 };
 
 const nextModeContent = {
-  canvas: 'SWITCH TO GL',
-  gl: 'SWITCH TO SVG',
+  canvas: 'SWITCH TO SVG',
   svg: 'SWITCH TO CANVAS'
 };
 
-const drawModes = ['canvas', 'gl', 'svg'];
+const drawModes = ['canvas', 'svg'];
 
 export default class Example extends React.Component {
   state = {
     drawMode: 0,
     data: randomData,
     colorType: 'typeA',
+    strokeWidth: 1,
+    showMarks: true,
     value: false
   }
 
   render() {
-    const {drawMode, data, colorType} = this.state;
-    const markSeriesProps = {
+    const {drawMode, data, colorType, strokeWidth, value, showMarks} = this.state;
+    const lineSeriesProps = {
       animation: true,
       className: 'mark-series-example',
       sizeRange: [5, 15],
-      seriesId: 'my-example-scatterplot',
+      color: colorType === 'typeA' ? '#0D676C' : '#B52F93',
       colorRange: colorRanges[colorType],
       opacityType: 'literal',
+      strokeWidth,
       data,
-      onNearestXY: value => this.setState({value})
+      onNearestX: d => this.setState({value: d})
     };
+    const SVGComponent = showMarks ? LineMarkSeries : LineSeries;
+    const CanvasComponent = showMarks ? LineMarkSeriesCanvas : LineSeriesCanvas;
 
     const mode = drawModes[drawMode];
     return (
-      <div className="scatterplot-gl-wrapper">
-        <div className="scatterplot-gl-example-controls">
-          <div>{`MODE: ${mode}`}</div>
+      <div className="canvas-wrapper">
+        <div className="canvas-example-controls">
+          <div> {`Mode: ${mode}`} </div>
           <ShowcaseButton
-            onClick={() => this.setState({drawMode: (drawMode + 1) % 3})}
+            onClick={() => this.setState({drawMode: (drawMode + 1) % 2})}
             buttonContent={nextModeContent[mode]} />
+          <ShowcaseButton
+            onClick={() => this.setState({showMarks: !showMarks})}
+            buttonContent={showMarks ? 'HIDE MARKS' : 'SHOW MARKS'} />
           <ShowcaseButton
             onClick={() => this.setState({data: getRandomData()})}
             buttonContent={'UPDATE DATA'} />
           <ShowcaseButton
             onClick={() => this.setState({colorType: nextType[colorType]})}
-            buttonContent={'UPDATE COLOR'} />
+            buttonContent={'TOGGLE COLOR'} />
+          <ShowcaseButton
+            onClick={() => this.setState({strokeWidth: strokeWidth === 1 ? 2 : 1})}
+            buttonContent={'TOGGLE STROKEWIDTH'} />
         </div>
         <XYPlot
           width={600}
@@ -104,16 +115,11 @@ export default class Example extends React.Component {
           <HorizontalGridLines />
           <XAxis />
           <YAxis />
-          {mode === 'gl' &&
-            <MarkSeriesGL {...markSeriesProps} seriesId="my-example-scatterplot"/>}
           {mode === 'canvas' &&
-            <MarkSeriesCanvas {...markSeriesProps}/>}
+            <CanvasComponent {...lineSeriesProps}/>}
           {mode === 'svg' &&
-            <MarkSeries {...markSeriesProps}/>}
-          {this.state.value ?
-            <Hint value={this.state.value}/> :
-            null
-          }
+            <SVGComponent {...lineSeriesProps}/>}
+          {value && <Crosshair values={[value]} />}
         </XYPlot>
       </div>
     );
