+++ src/index.js
@@ -21,7 +21,6 @@
 export AbstractSeries from 'plot/series/abstract-series';
 export LineSeries from 'plot/series/line-series';
 export LineSeriesCanvas from 'plot/series/line-series-canvas';
-export LineSeriesGL from 'plot/series/line-series-gl';
 export HorizontalBarSeries from 'plot/series/horizontal-bar-series';
 export HorizontalBarSeriesCanvas from 'plot/series/horizontal-bar-series-canvas';
 export VerticalBarSeries from 'plot/series/vertical-bar-series';
@@ -35,14 +34,12 @@ export PolygonSeries from 'plot/series/polygon-series';
 export RectSeries from 'plot/series/rect-series';
 export RectSeriesCanvas from 'plot/series/rect-series-canvas';
 export MarkSeries from 'plot/series/mark-series';
-export MarkSeriesGL from 'plot/series/mark-series-gl';
 export MarkSeriesCanvas from 'plot/series/mark-series-canvas';
 export HeatmapSeries from 'plot/series/heatmap-series';
 export AreaSeries from 'plot/series/area-series';
 export ArcSeries from 'plot/series/arc-series';
 export LineMarkSeries from 'plot/series/line-mark-series';
 export LineMarkSeriesCanvas from 'plot/series/line-mark-series-canvas';
-export LineMarkSeriesGL from 'plot/series/line-mark-series-gl';
 export Hint from 'plot/hint';
 export Crosshair from 'plot/crosshair';
 export XYPlot from 'plot/xy-plot';
