+++ src/plot/xy-plot.js
@@ -28,7 +28,6 @@ import {getInnerDimensions, MarginPropType} from 'utils/chart-utils';
 import {AnimationPropType} from 'animation';
 import {CONTINUOUS_COLOR_RANGE, SIZE_RANGE, OPACITY_TYPE} from 'theme';
 
-import DeckGLWrapper from './series/deck-gl-wrapper';
 import CanvasWrapper from './series/canvas-wrapper';
 
 const ATTRIBUTES = [
@@ -315,33 +314,6 @@ class XYPlot extends React.Component {
     });
   }
 
-  renderDeckGlComponents(components, props) {
-    const componentsToRender = components.filter(c => c && !c.type.requiresSVG && c.type.isDeckGL);
-    if (componentsToRender.length === 0) {
-      return null;
-    }
-    const animation = componentsToRender.some(layer => layer.props.animation);
-    const {
-      marginLeft,
-      marginTop,
-      marginBottom,
-      marginRight,
-      innerHeight,
-      innerWidth
-    } = componentsToRender[0].props;
-    return (<DeckGLWrapper {...{
-      animation,
-      marginLeft,
-      marginTop,
-      marginBottom,
-      marginRight,
-      innerHeight,
-      innerWidth
-    }}>
-      {componentsToRender}
-    </DeckGLWrapper>);
-  }
-
   renderCanvasComponents(components, props) {
     const componentsToRender = components.filter(c => c && !c.type.requiresSVG && c.type.isCanvas);
 
@@ -405,10 +377,9 @@ class XYPlot extends React.Component {
           onMouseMove={this._mouseMoveHandler}
           onMouseLeave={this._mouseLeaveHandler}
           onMouseEnter={this._mouseEnterHandler}>
-          {components.filter(c => c && c.type.requiresSVG && !c.type.isDeckGL)}
+          {components.filter(c => c && c.type.requiresSVG)}
         </svg>
         {this.renderCanvasComponents(components, this.props)}
-        {this.renderDeckGlComponents(components, this.props)}
         {components.filter(c => c && !c.type.requiresSVG)}
       </div>
     );
