+++ src/styles/examples.scss
@@ -383,11 +383,11 @@ article {
   }
 }
 
-.scatterplot-gl-example-controls {
+.canvas-example-controls {
   display: flex;
 }
 
-.scatterplot-gl-wrapper {
+.canvas-wrapper {
   align-items: center;
   display: flex;
   flex-direction: column;
